requirements_2.txt
absl-py==1.1.0
addict==2.4.0
basicsr==1.3.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
facexlib==0.2.4
ffmpeg==1.4
ffmpeg-python==0.2.0
filterpy==1.4.5
fonttools==4.34.4
future==0.18.2
gfpgan==1.3.2
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
kiwisolver==1.4.3
llvmlite==0.38.1
lmdb==1.3.0
Markdown==3.3.7
matplotlib==3.5.2
networkx==2.8.4
numba==0.55.2
numpy==1.23.0
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
tb-nightly==2.10.0a20220707
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.5.4
torch==1.12.0+cpu
torchaudio==0.12.0+cpu
torchvision==0.13.0+cpu
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
yapf==0.32.0
zipp==3.8.0
